refactor: 로그인 코드스타일 통일
